pyproject.toml
[project]
name = "1box-dashboard"
version = "0.1.0"
description = "Marketing analytics dashboard for 1BOX Self-Storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.133.0",
    "google-analytics-data>=0.20.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sqlalchemy>=2.0.47",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.2",
]
